+++ test/integration/IntegrationTest.cpp
@@ -117,6 +117,7 @@ BOOST_AUTO_TEST_CASE(Logging)               { runReadDiff("Logging", "%S %m"); }
 BOOST_AUTO_TEST_CASE(LoggingFundamentals)   { runReadDiff("LoggingFundamentals", "%m"); }
 BOOST_AUTO_TEST_CASE(LoggingContainers)     { runReadDiff("LoggingContainers", "%m"); }
 BOOST_AUTO_TEST_CASE(LoggingStrings)        { runReadDiff("LoggingStrings", "%m"); }
+BOOST_AUTO_TEST_CASE(LoggingCStrings)       { runReadDiff("LoggingCStrings", "%m"); }
 BOOST_AUTO_TEST_CASE(LoggingPointers)       { runReadDiff("LoggingPointers", "%m"); }
 BOOST_AUTO_TEST_CASE(LoggingTuples)         { runReadDiff("LoggingTuples", "%m"); }
 BOOST_AUTO_TEST_CASE(LoggingEnums)          { runReadDiff("LoggingEnums", "%m"); }